separate timeline api
